requirements.txt
discord.py==2.0.1
PyNaCl==1.5.0
requests==2.28.1
youtube-dl==2021.12.17
